Refactor header files organization
Our header files are mixing declarations intended for the user facing
API and declaration for internal code. This was creating problems with
adding new code and exposing functionality to the users in a clean and
error-free way.

This commit refactors the organization of the header files by splitting
the user facing API from internal data-structures and functions. User
facing API is now located under `src/include` whereas header files for
the runtime itself are located in same folder as the corresponding
source code file.
