+++ src/error.h
@@ -1,0 +1,6 @@
+#ifndef __ERROR_H__
+#define __ERROR_H__
+
+#include "include/error.h"
+
+#endif /* __ERROR_H__ */
