+++ src/include/ops/noop.h
@@ -1,0 +1,16 @@
+#ifndef __VACCEL_NOOP_H__
+#define __VACCEL_NOOP_H__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct vaccel_session;
+
+int vaccel_noop(struct vaccel_session *sess);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* __VACCEL_NOOP_H__ */
