+++ src/include/vaccel.h
@@ -1,16 +1,15 @@
 #ifndef __VACCEL_H__
 #define __VACCEL_H__
 
-#include <stdint.h>
-#include <stddef.h>
-
 #include "error.h"
-
+#include "log.h"
+#include "plugin.h"
+#include "session.h"
+#include "ops/vaccel_ops.h"
 #include "ops/blas.h"
 #include "ops/exec.h"
 #include "ops/genop.h"
 #include "ops/image.h"
 #include "ops/noop.h"
-#include "ops/vaccel_ops.h"
 
 #endif /* __VACCEL_H__ */
