+++ src/ops/noop.c
@@ -1,9 +1,11 @@
 #include "noop.h"
-#include "common.h"
+#include "error.h"
 #include "plugin.h"
 #include "log.h"
 #include "vaccel_ops.h"
 #include "genop.h"
+
+#include "session.h"
 
 int vaccel_noop(struct vaccel_session *sess)
 {
