+++ src/ops/noop.h
@@ -1,12 +1,13 @@
-#ifndef __VACCEL_NOOP_H__
-#define __VACCEL_NOOP_H__
+#ifndef __NOOP_H__
+#define __NOOP_H__
+
+#include "include/ops/noop.h"
 
 struct vaccel_session;
 struct vaccel_arg;
 
-int vaccel_noop(struct vaccel_session *sess);
 int vaccel_noop_unpack(struct vaccel_session *sess,
 		struct vaccel_arg *read, int nr_read,
 		struct vaccel_arg *write, int nr_write);
 
-#endif /* __VACCEL_NOOP_H__ */
+#endif /* __NOOP_H__ */
