+++ src/session.c
@@ -1,4 +1,4 @@
-#include "vaccel.h"
+#include "session.h"
 #include "plugin.h"
 #include "log.h"
