+++ src/session.h
@@ -1,5 +1,7 @@
-#ifndef __VACCEL_SESSION_H__
-#define __VACCEL_SESSION_H__
+#ifndef __SESSION_H__
+#define __SESSION_H__
+
+#include "include/session.h"
 
 int sessions_bootstrap(void);
